requirements.txt
fastapi
uvicorn
SQLAlchemy<2.0
passlib
bcrypt
python-jose
python-multipart